setup minimum and maximum lane width protocol
If the lane width jumps up above 4m assume that the left lane is incorrect and bring it to be closer so that the lane width is exactly 4m
If the lane width is smaller than 2m which could be if there is a bicycle lane being detected that is in the middle of our lane add the extra distance to the right lane and assume the middle lane (left lane) to be correct. This should be swapped if is_rhd is active.